+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 Framesync adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.1.1] 2017-11-08
+
+### Fixed
+
+- Automatically exporting declaration file.
+
+## [3.1.0] 2017-11-08
+
+### Added
+
+- Added optional `true` flag to `schedule`. This will schedule a job to run at the end of the current frame step.
+
 ## [3.0.0] 2017-08-28
 
 - `currentFrameTimestamp` becomes `currentFrameTime` for symmetry with `currentTime`.